requirements.txt
fastapi==0.116.1
pymongo==4.13.2
uvicorn==0.35.0
python-dotenv==1.1.1
firebase-admin==7.0.0